+++ ebl/corpus/infrastructure/mongo_text_repository_modify.py
@@ -1,0 +1,70 @@
+import pymongo
+from ebl.corpus.application.schemas import (
+    ChapterSchema,
+    TextSchema,
+)
+from ebl.corpus.domain.chapter import Chapter, ChapterId
+from ebl.corpus.domain.text import Text
+from ebl.corpus.infrastructure.queries import (
+    chapter_id_query,
+)
+from ebl.corpus.infrastructure.mongo_text_repository_base import MongoTextRepositoryBase
+
+
+class MongoTextRepositoryModify(MongoTextRepositoryBase):
+    def create_indexes(self) -> None:
+        self._texts.create_index(
+            [
+                ("genre", pymongo.ASCENDING),
+                ("category", pymongo.ASCENDING),
+                ("index", pymongo.ASCENDING),
+            ],
+            unique=True,
+        )
+        self._chapters.create_index(
+            [
+                ("textId.genre", pymongo.ASCENDING),
+                ("textId.category", pymongo.ASCENDING),
+                ("textId.index", pymongo.ASCENDING),
+            ]
+        )
+        self._chapters.create_index(
+            [
+                ("textId.genre", pymongo.ASCENDING),
+                ("textId.category", pymongo.ASCENDING),
+                ("textId.index", pymongo.ASCENDING),
+                ("order", pymongo.ASCENDING),
+            ]
+        )
+        self._chapters.create_index(
+            [
+                ("textId.genre", pymongo.ASCENDING),
+                ("textId.category", pymongo.ASCENDING),
+                ("textId.index", pymongo.ASCENDING),
+                ("stage", pymongo.ASCENDING),
+                ("name", pymongo.ASCENDING),
+            ],
+            unique=True,
+        )
+
+    def create(self, text: Text) -> None:
+        self._texts.insert_one(TextSchema(exclude=["chapters"]).dump(text))
+
+    def create_chapter(self, chapter: Chapter) -> None:
+        self._chapters.insert_one(ChapterSchema().dump(chapter))
+
+    def update(self, id_: ChapterId, chapter: Chapter) -> None:
+        self._chapters.update_one(
+            chapter_id_query(id_),
+            {
+                "$set": ChapterSchema(
+                    only=[
+                        "manuscripts",
+                        "uncertain_fragments",
+                        "lines",
+                        "signs",
+                        "parser_version",
+                    ]
+                ).dump(chapter)
+            },
+        )
